+++ minitrade/trader/schema/0.1.36.sql
@@ -1,0 +1,6 @@
+CREATE TABLE IF NOT EXISTS "UnionQuoteSourceConfig" (
+    "name" TEXT NOT NULL,
+    "config" JSON NOT NULL,
+    "update_time" DATETIME NOT NULL,
+    PRIMARY KEY("name")
+);