+++ minitrade/trader/schema/consolidated.sql
@@ -244,16 +244,23 @@ CREATE TABLE IF NOT EXISTS "TaskLog" (
 
 CREATE TABLE IF NOT EXISTS "CboeFuturesCache" (
 	"hash" TEXT NOT NULL,
-	'product_display' TEXT,
-	'expire_date' TEXT,
-	'contract_dt' TEXT,
-	'futures_root' TEXT,
-	'duration_type' TEXT,
+	"product_display" TEXT,
+	"expire_date" TEXT,
+	"contract_dt" TEXT,
+	"futures_root" TEXT,
+	"duration_type" TEXT,
 	"freeze" BOOLEAN NOT NULL,
 	"update_time" DATETIME,
 	"path" TEXT,
 	"data" TEXT,
 	PRIMARY KEY("hash")
+);
+
+CREATE TABLE IF NOT EXISTS "UnionQuoteSourceConfig" (
+	"name" TEXT NOT NULL,
+	"config" JSON NOT NULL,
+	"update_time" DATETIME NOT NULL,
+	PRIMARY KEY("name")
 );
 
 COMMIT;